+++ src/compas_fab/datastructures/__init__.py
@@ -1,0 +1,35 @@
+"""
+********************************************************************************
+compas_fab.datastructures
+********************************************************************************
+
+.. currentmodule:: compas_fab.datastructures
+
+PartialOrder
+-----
+
+.. autosummary::
+    :toctree: generated/
+    :nosignatures:
+
+    Action
+    DependencyIdException
+    IntegerIdGenerator
+    PartialOrder
+
+"""
+
+from .partial_order import (
+    Action,
+    DependencyIdException,
+    IntegerIdGenerator,
+    PartialOrder
+)
+
+
+__all__ = [
+    'Action',
+    'DependencyIdException',
+    'IntegerIdGenerator',
+    'PartialOrder',
+]