+++ CHANGELOG.rst
@@ -67,6 +67,7 @@ Unreleased
 * Grasshopper components now also for Mac
 * Added support for MoveIt on ROS Noetic
 * Added support for Python 3.9
+* In ``compas.datastructures``, added ``Plan``, ``Action`` and ``IntegerIdGenerator``
 
 **Changed**